% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/emotion_analysis.R
\name{emotion_analysis}
\alias{emotion_analysis}
\title{Assessment of emotional valence and arousal}
\usage{
emotion_analysis(data)
}
\arguments{
\item{data}{Data frame containing ECG, EEG, GPS and heart rate data.}
}
\value{
The function returns the input data frame complemented with columns:
 \itemize{
   \item \code{timecest}: the time vector in \code{POSIXct} format (converted from the \code{time} column),
   \item \code{fp1_fp2_beta_alpha_avg}: average values of beta-alpha ratios of channels FP1 and FP2,
   \item \code{arousal}: factorized arousal classes averaged for channels FP1 and FP2,
   \item \code{valence}: factorized valence classes for channels FP1 and FP2,
   \item \code{arousalFP1}: factorized arousal classes for channel FP1,
   \item \code{arousalFP2}: factorized arousal classes for channel FP1. }
}
\description{
This function uses the EEG data collected with Muse and assesses emotional valence and arousal based on asymmetry analysis.
The EEG data contains EEG signal in microvolts and alpha and beta absolute band powers (in Bels!).
Emotional valence is calcluated based on the ratio of alpha power between right and left EEG chanels.
Emotional arousal is calculated based on the mean value of beta to alpha ratios of left and right EEG channels.
}
\keyword{EEG}
\keyword{EEG,}
\keyword{alpha,}
\keyword{arousal,}
\keyword{assymetry}
\keyword{beta,}
\keyword{valence,}

