% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_eeg.R
\name{read_eeg}
\alias{read_eeg}
\title{Read Muse EEG data file}
\usage{
read_eeg(NAME = "filename")
}
\arguments{
\item{NAME}{The input parameter is the string path to the selected ECG file (csv format).}
}
\value{
Function returns the eeg data frame.
 The returned EEG data frame contains the following columns:
   - time: vector of unix timestamps
   - type: indication of wether the signal is EEG, or only alpha or beta band
   - TP9, FP1, FP2, TP10: EEG channels
   - fp1_fp2: ratio of alpha and beta band data of channels FP1 and FP2 (in decibels).
   - fp1_beta_alpha: ratio between beta and alpha band power of channel FP1 (in decibels).
   - fp2_beta_alpha: ratio between beta and alpha band power of channel FP2 (in decibels).
}
\description{
This function reads raw EEG data in csv format acquired by MUSE hardware and sorts it into a data frame.
The input parameter is the string path to the selected EEG file (csv format).
Function returns the data frame with EEG data.
The appropriate file must be converted from .muse format to .csv format using muse.player.
}
\examples{
eeg <- read_eeg(system.file("extdata", "eeg_sample.csv", package="museEEG"))
}
\keyword{EEG}
\keyword{Muse}
\keyword{data,}
\keyword{read,}

